Refactor: Migrate Calculator State to BigDecimal and Improve Architecture
This commit refactors the calculator's internal state management to use `BigDecimal` as the primary data type for inputs, replacing String-based handling. It also streamlines the UI architecture by merging `CalculatorContent` into `CalculatorLayout` and improving the `ViewModel` integration.

- **`CalculateState.kt`:**
    - Changed `input` type from `String` to `BigDecimal`.
    - Removed `derivedStateOf` usage, replacing it with direct property accessors (`pendingMemory`, `inputString`) for better performance and simplicity.

- **`CalculatorBrain.kt`:**
    - Introduced `onNumberAction` to specifically handle numeric and decimal inputs, separating them from general actions.
    - Added a `setup(initialValue)` method to initialize the calculator state once.
    - Updated all arithmetic operations (`onAdd`, `onSubtract`, etc.) to perform calculations directly on `BigDecimal` values.
    - Refactored `onBackSpace` and `onToggleSign` to handle `BigDecimal` state.

- **`CalculatorContent.kt` & `CalculatorScreen.kt`:**
    - Removed the redundant `CalculatorContent` wrapper.
    - Updated `CalculatorLayout` to accept a `Modifier` and use `state.inputString` for display.
    - Refactored `CalculatorAlertDialog` to manage its own `CalculatorBrain` instance and observation logic.
    - Added "Speichern" (Save) and "Abbrechen" (Cancel) button logic to the dialog.

- **`Numpad.kt` & Tests:**
    - Updated `Numpad` and `CalculatorBrainTest` to use the renamed `onNumberAction` method.
    - Updated preview blocks and tests to reflect the change from String to `BigDecimal` in `UICalculateState`.

+++ calculator/src/main/java/com/gerwalex/calculator/common/ActionButton.kt
@@ -81,7 +81,7 @@ fun ActionButton(
 @Composable
 private fun CalculatorActionButtonDialog() {
     val state = UICalculateState(
-            input = "123",
+        input = "123".toBigDecimal(),
             pendingValue = BigDecimal(456),
             pendingOperation = ActionButtonType.Add
         )
